refactor: extract composer submit from AgentWorkspace
