+++ src/components/agent/agent-workspace-models.ts
@@ -1,5 +1,6 @@
 import type { HermesAttachmentState } from "../../lib/hermes-image-attach";
 import type { ImportedHermesFile } from "../../lib/tauri";
+import type { HermesSessionDispatchReservation } from "../../lib/hermes-session-dispatch-mutex";
 
 export type AgentAttachment = ImportedHermesFile & {
   id: string;
@@ -14,4 +15,20 @@ export type AgentAttachment = ImportedHermesFile & {
    * or failed. Carries file refs only, never the image bytes. Files stay
    * `imported` (they only ride along as a path in the prompt). */
   attach: HermesAttachmentState;
+};
+
+export type ImageSafeModeConsentChoice =
+  | { action: "keep"; dontAskAgain: boolean }
+  | { action: "turnOff"; dontAskAgain: boolean }
+  | { action: "dismiss" };
+
+export type ImageSafeModeConsentRequest = {
+  variant: "slash" | "agent" | "video-slash";
+  ownerDispatchReservation?: HermesSessionDispatchReservation;
+  resolve: (choice: ImageSafeModeConsentChoice) => void;
+};
+
+export type ImageSafeModeConsentEventPayload = {
+  source?: string;
+  prompt?: string;
 };
